affiliate-site/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
SQLAlchemy==2.0.35
passlib[bcrypt]==1.7.4
